requirements.txt
olefile==0.44
Pillow==4.3.0
py==1.4.34
pytest==3.2.3
